pyproject.toml
[project]
name = "genetic-discrimination-policy-econ"
version = "0.2.0"
description = "JAX-first scaffold for integrated economic evaluation of genetic discrimination policy options"
requires-python = ">=3.10"
dependencies = [
  # Core numerical stack
  "jax>=0.4.25",
  "jaxlib>=0.4.25",
  "numpy>=1.26",
  "polars>=0.20.0",  # Faster than pandas
  
  # Probabilistic programming
  "numpyro>=0.15.0",
  "optax>=0.2.2",
  "jaxopt>=0.8.3",
  "distrax>=0.1.5",
  "equinox>=0.11.3",
  "chex>=0.1.86",
  "jaxtyping>=0.2.25",
  
  # Config management
  "pyyaml>=6.0",
  "pydantic>=2.6",
  "hydra-core>=1.3",
  
  # Game theory / economics
  "nashpy>=0.0.32",
  "quantecon>=0.8.0",
  "networkx>=3.0",
  
  # CLI and output
  "typer>=0.9.0",
  "rich>=13.7.0",
  "tqdm>=4.66.0",
  
  # Reporting
  "matplotlib>=3.8",
  "tabulate>=0.9",
  "python-docx>=1.1",
  "reportlab>=4.0",
]

[project.optional-dependencies]
dev = [
  "uv>=0.1.0",  # Ultra-fast package installer
  "nox>=2024.3.2",  # Environment management
  "loguru>=0.7.2",  # Modern logging
  "pytest>=8.0",
  "pytest-cov>=4.0",
  "pytest-xdist>=3.5",  # Parallel testing
  "pytest-mock>=3.12",
  "pytest-timeout>=2.2",
  "hypothesis>=6.97",  # Property-based testing
  "mutmut>=2.4",  # Mutation testing
  "ruff>=0.3.0",
  "pyright>=1.1.350",  # Faster than mypy
  "pre-commit>=3.6",
  "ipython>=8.20",
  
  # Runtime type checking
  "beartype>=0.18.0",  # O(1) runtime type checking
  
  # Data classes
  "attrs>=23.2.0",  # JAX-compatible data classes
  
  # Serialization
  "msgspec>=0.18.0",  # Fast serialization/validation
]

profiling = [
  # CPU Profilers
  "py-spy>=0.3.14",  # Sampling profiler (works with JAX)
  "pyinstrument>=4.6",  # Call stack profiler
  "scalene>=1.5.40",  # CPU+GPU+memory profiler (SOTA)
  "line_profiler>=4.1",  # Line-by-line timing
  
  # Memory Profilers
  "memray>=1.11",  # Memory profiler with flame graphs
  "tracemalloc-debug>=0.1",  # Python memory tracing
  
  # Visualization
  "snakeviz>=2.2.0",  # cProfile viewer
  "speedscope>=0.0.1",  # Interactive flame graphs
]

coverage = [
  "coverage>=7.4",  # Core coverage
  "coverage-badge>=1.1",  # Coverage badges
  "diff-cover>=8.0",  # Coverage on diffs
  "pytest-cov>=4.0",
  "covdefaults>=2.3",  # Sensible coverage defaults
]

validation = [
  "bibtex-tidy>=1.0",
  "bibtexparser>=1.4",
  "requests>=2.31",
  "habanero>=1.2",
]

workflow = [
  "snakemake>=8.0",
  "dvc>=3.0",
]

[project.scripts]
run-model = "scripts.run_model:main"
run-stress-tests = "scripts.run_stress_tests:main"
run-voi = "scripts.run_voi:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
  "E",    # pycodestyle errors
  "W",    # pycodestyle warnings
  "F",    # pyflakes
  "I",    # isort
  "N",    # pep8-naming
  "UP",   # pyupgrade
  "ANN",  # flake8-annotations
  "ASYNC",# flake8-async
  "S",    # flake8-bandit
  "B",    # flake8-bugbear
  "A",    # flake8-builtins
  "COM",  # flake8-commas
  "C4",   # flake8-comprehensions
  "DTZ",  # flake8-datetimez
  "T10",  # flake8-debugger
  "EM",   # flake8-errmsg
  "EXE",  # flake8-executable
  "FA",   # flake8-future-annotations
  "ISC",  # flake8-implicit-str-concat
  "ICN",  # flake8-import-conventions
  "G",    # flake8-logging-format
  "INP",  # flake8-no-pep420
  "PIE",  # flake8-pie
  "T20",  # flake8-print
  "PYI",  # flake8-pyi
  "PT",   # flake8-pytest-style
  "Q",    # flake8-quotes
  "RSE",  # flake8-raise
  "RET",  # flake8-return
  "SLF",  # flake8-self
  "SIM",  # flake8-simplify
  "TID",  # flake8-tidy-imports
  "TCH",  # flake8-type-checking
  "INT",  # flake8-gettext
  "ARG",  # flake8-unused-arguments
  "PTH",  # flake8-use-pathlib
  "TD",   # flake8-todos
  "ERA",  # eradicate
  "PD",   # pandas-vet
  "PGH",  # pygrep-hooks
  "PL",   # pylint
  "TRY",  # tryceratops
  "FLY",  # flynt
  "PERF", # perflint
  "RUF",  # ruff-specific rules
]
ignore = [
  "E501",   # line length (handled by formatter)
  "ANN101", # missing-type-self
  "ANN102", # missing-type-cls
  "TD002",  # missing-todo-author
  "TD003",  # missing-todo-link
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "S101", "PLR2004", "ARG", "FBT", "INP001"]
"scripts/*" = ["ANN", "T201", "INP001"]
"examples/*" = ["ANN", "T201", "S101", "INP001"]
"streamlit_app/*" = ["ANN", "T201"]

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
exclude = ["**/tests/**", "**/scripts/**", ".venv", "outputs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=xml -n auto"
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "property: property-based tests with hypothesis",
  "mutation: mutation tests",
]
timeout = 300  # 5 minute default timeout

[tool.hypothesis]
database = ".hypothesis/examples"
deadline = 500
max_examples = 100
suppress_health_check = ["too_slow"]
report_multiple_bugs = true

[tool.mutmut]
runner = "pytest -x"
tests_dir = "tests/"
paths_to_mutate = "src/"
backup = false
use_patch = true
dict_synonyms = ["Partial", "NamedTuple"]

[tool.coverage.run]
source = ["src"]
branch = true
parallel = true
concurrency = ["multiprocessing", "thread"]
sigterm = true
plugins = ["covdefaults"]

[tool.coverage.paths]
source = [
  "src/",
  "*/site-packages/",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "@abstractmethod",
  "class .*\\bProtocol\\):",
]

show_missing = true
precision = 2
fail_under = 85.0

[tool.coverage.html]
directory = "outputs/coverage_html"
title = "Coverage Report"

[tool.coverage.xml]
output = "coverage.xml"

[tool.diff_cover]
compare_branch = "origin/main"
diff_range_notation = ".."

[tool.hydra]
output_subdir = ".hydra"
run.dir = "outputs/${now:%Y-%m-%d}/${now:%H-%M-%S}"
